Refactored: Datadog::Diagnostics::EnvironmentLogger --> Datadog::Core::Diagnostics::EnvironmentLogger
