+++ lib/ddtrace/workers/trace_writer.rb
@@ -53,7 +53,7 @@ module Datadog
         end
       end
 
-      # TODO: Register `Datadog::Diagnostics::EnvironmentLogger.log!`
+      # TODO: Register `Datadog::Core::Diagnostics::EnvironmentLogger.log!`
       # TODO: as a flush_completed subscriber when the `TraceWriter`
       # TODO: instantiation code is implemented.
       def flush_completed
