+++ spec/ddtrace/configuration/components_spec.rb
@@ -771,7 +771,7 @@ RSpec.describe Datadog::Configuration::Components do
       let(:responses) { [double('response')] }
 
       it 'invokes the environment logger with responses' do
-        expect(Datadog::Diagnostics::EnvironmentLogger).to receive(:log!).with(responses)
+        expect(Datadog::Core::Diagnostics::EnvironmentLogger).to receive(:log!).with(responses)
         call
       end
     end
